# engageiu/openapi.yaml
openapi: "3.0.3"
info:
  title: EngageIU API
  version: "1.0.0"
  description: >
    QR/code-based campus event attendance leaderboard system for Indiana University.
    Students earn points by entering check-in codes at IU events.
    Built for Luddy Hacks — Dynamic Leaderboard/Ranking System case.
  contact:
    name: EngageIU Team
    email: iuhacks@iu.edu
servers:
  - url: http://localhost:8000
    description: Local development server
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Alex Johnson"
        iu_username:
          type: string
          example: "alexj"
        campus:
          type: string
          example: "IU Bloomington"
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
          example: 1
        name:
          type: string
          example: "Alex Johnson"
        iu_username:
          type: string
          example: "alexj"
        campus:
          type: string
          example: "IU Bloomington"
        total_points:
          type: integer
          example: 60
        events_attended:
          type: integer
          example: 3
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        student:
          $ref: '#/components/schemas/Student'
        event:
          type: object
          properties:
            id:
              type: integer
            title:
              type: string
            campus:
              type: string
        points_earned:
          type: integer
        checked_in_at:
          type: string
          format: date-time
    InfoBasic:
      type: object
      properties:
        total_students:
          type: integer
          example: 150
        total_events:
          type: integer
          example: 12
        students_with_points_this_week:
          type: integer
          example: 47
        mean:
          type: number
          example: 28.4
        median:
          type: number
          example: 25.0
        q1:
          type: number
          example: 15.0
        q3:
          type: number
          example: 40.0
        min:
          type: number
          example: 10.0
        max:
          type: number
          example: 75.0
    InfoExtended:
      allOf:
        - $ref: '#/components/schemas/InfoBasic'
        - type: object
          properties:
            std_deviation:
              type: number
              example: 14.7
            percentile_ranks:
              type: array
              items:
                type: object
                properties:
                  score:
                    type: number
                  percentile_rank:
                    type: number
            score_distribution:
              type: array
              items:
                type: object
                properties:
                  range:
                    type: string
                    example: "10–23"
                  count:
                    type: integer
                  percentage:
                    type: number
            top_campus:
              type: string
              example: "IU Bloomington"
            most_attended_event:
              type: string
              example: "Hoosier Innovation Summit"
            weekly_growth_rate_pct:
              type: number
              nullable: true
              example: 23.5
    PerformanceStats:
      type: object
      properties:
        endpoints:
          type: array
          items:
            type: object
            properties:
              endpoint:
                type: string
                example: "/leaderboard"
              method:
                type: string
                example: "GET"
              avg_ms:
                type: number
                example: 12.4
              min_ms:
                type: number
                example: 8.1
              max_ms:
                type: number
                example: 45.2
              call_count:
                type: integer
                example: 38
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Not found"
...
paths:
...
  /add:
    post:
      tags: [leaderboard]
      summary: Add a leaderboard entry
      description: >
        Admin-only. Validates the check-in code for the specified event,
        creates the student if they don't exist, and records attendance.
        Returns the student's updated rank and total points.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [iu_username, name, campus, event_id, check_in_code]
              properties:
                iu_username:
                  type: string
                  example: "jdoe"
                name:
                  type: string
                  example: "Jane Doe"
                campus:
                  type: string
                  example: "IU Bloomington"
                event_id:
                  type: integer
                  example: 1
                check_in_code:
                  type: string
                  example: "INNOV2025"
      responses:
        "200":
          description: Entry added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  student:
                    $ref: '#/components/schemas/Student'
                  points_earned:
                    type: integer
                    example: 20
                  new_total:
                    type: integer
                    example: 35
                  rank:
                    type: integer
                    example: 3
        "400":
          description: Invalid check-in code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized — admin token required
        "404":
          description: Event not found
        "409":
          description: Student already checked in to this event
  /remove:
    delete:
      tags: [leaderboard]
      summary: Remove a leaderboard entry
      description: Admin-only. Deletes the attendance record for a student/event pair.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [iu_username, event_id]
              properties:
                iu_username:
                  type: string
                  example: "jdoe"
                event_id:
                  type: integer
                  example: 1
      responses:
        "200":
          description: Entry removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "removed"
                  student:
                    $ref: '#/components/schemas/Student'
        "401":
          description: Unauthorized
        "404":
          description: Student or attendance record not found
  /leaderboard:
    get:
      tags: [leaderboard]
      summary: Top 10 weekly leaderboard
      description: >
        Returns the top 10 students ranked by total points earned during the
        current week (Sunday–Saturday). Supports campus filtering and an HTML
        table output format. Week boundary can be shifted with the ?week param.
      parameters:
        - in: query
          name: campus
          schema:
            type: string
          description: Filter by IU campus name
          example: "IU Bloomington"
        - in: query
          name: week
          schema:
            type: string
            format: date
          description: ISO date within the desired week (defaults to current week)
          example: "2025-04-14"
        - in: query
          name: format
          schema:
            type: string
            enum: [json, html]
          description: "Response format: json (default) or html table"
      responses:
        "200":
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  week_start:
                    type: string
                    format: date-time
                  week_end:
                    type: string
                    format: date-time
                  campus_filter:
                    type: string
                    nullable: true
                  leaderboard:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
            text/html:
              schema:
                type: string
  /info:
    get:
      tags: [leaderboard]
      summary: Score statistics
      description: >
        Returns quantitative statistics for all weekly scores.
        Public: mean, median, Q1, Q3, min, max, counts.
        Admin (with bearer token): additionally returns std_deviation,
        percentile_ranks, score_distribution, top_campus,
        most_attended_event, and weekly_growth_rate_pct.
      security:
        - bearerAuth: []
        - {}
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/InfoBasic'
                  - $ref: '#/components/schemas/InfoExtended'
  /performance:
    get:
      tags: [leaderboard]
      summary: Endpoint performance statistics
      description: >
        Admin-only. Returns average, min, max response time and call count
        per endpoint, sorted by average response time descending.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceStats'
        "401":
          description: Unauthorized
  /history:
    get:
      tags: [leaderboard]
      summary: Score submission history (grad requirement)
      description: >
        Admin-only. Paginated log of all attendance/score submissions
        with timestamps. Supports filtering by date range and IU username.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          description: Filter submissions on or after this date (YYYY-MM-DD)
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          description: Filter submissions on or before this date (YYYY-MM-DD)
        - in: query
          name: iu_username
          schema:
            type: string
          description: Partial match on IU username
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated history
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceRecord'
        "401":
          description: Unauthorized

  # ── EVENTS ───────────────────────────────────────────────────────────────
...